Improve load speed by restoring cells from file
Cells are now saved along with the map's objects and restored from the file rather than being regenerated on load, which reduces load time by up to 75%. In support of this, cells are now auto-deleted when they contain no objects.

Some refactoring has been done to prepare for the creation of a core loader that can be incorporated directly into game code without the bulk of the full editor.